Create protocol handler service
