+++ src/app/components/project-map/context-menu/actions/console-device-action-browser/console-device-action-browser.component.ts
@@ -4,6 +4,8 @@ import { Node } from '../../../../../cartography/models/node';
 import { Server } from '../../../../../models/server';
 import { NodeService } from '../../../../../services/node.service';
 import { ToasterService } from '../../../../../services/toaster.service';
+import { ProtocolHandlerService } from '../../../../../services/protocol-handler.service';
+
 import * as ipaddr from 'ipaddr.js';
 
 @Component({
@@ -14,38 +16,13 @@ export class ConsoleDeviceActionBrowserComponent {
   @Input() server: Server;
   @Input() node: Node;
 
-  constructor(private toasterService: ToasterService, private nodeService: NodeService, private deviceService: DeviceDetectorService) {}
+  constructor(private toasterService: ToasterService, private nodeService: NodeService, private protocolHandlerService: ProtocolHandlerService) {}
 
   openConsole() {
     this.nodeService.getNode(this.server, this.node).subscribe((node: Node) => {
       this.node = node;
       this.startConsole();
     });
-  }
-
-  createHiddenIframe(target: Element, uri: string) {
-    const iframe = document.createElement("iframe");
-    iframe.src = uri;
-    iframe.id = "hiddenIframe";
-    iframe.style.display = "none";
-    target.appendChild(iframe);
-    return iframe;
-  }
-
-  openUriUsingFirefox(uri: string) {
-      var iframe = (document.querySelector("#hiddenIframe") as HTMLIFrameElement);
-
-      if (!iframe) {
-          iframe = this.createHiddenIframe(document.body, "about:blank");
-      }
-
-      try {
-          iframe.contentWindow.location.href = uri;
-      } catch (e) {
-          if (e.name === "NS_ERROR_UNKNOWN_PROTOCOL") {
-              this.toasterService.error('Protocol handler does not exist');
-          }
-      }
   }
 
   startConsole() {
@@ -59,8 +36,6 @@ export class ConsoleDeviceActionBrowserComponent {
       ) {
         this.node.console_host = this.server.host;
       }
-
-      const device = this.deviceService.getDeviceInfo();
 
       try {
         var uri;
@@ -81,13 +56,7 @@ export class ConsoleDeviceActionBrowserComponent {
           this.toasterService.error('Supported console types are: telnet, vnc, spice and spice+agent.');
         }
 
-        if (device.browser === "Firefox") {
-            // Use a hidden iframe otherwise Firefox will disconnect
-            // from the GNS3 controller websocket if we use location.assign()
-            this.openUriUsingFirefox(uri);
-        } else {
-            location.assign(uri);
-        }
+        this.protocolHandlerService.open(uri);
 
       } catch (e) {
           this.toasterService.error(e);
